pyproject.toml
[tool.poetry]
name = "python-django-drf"
version = "0.0.1"
description = ""
authors = ["Felipe Rodríguez"]
readme = "README.md"
packages = [{include = "python_django_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
yarl = "^1.9.2"
django-model-utils = "^4.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
ruff = "^0.0.280"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
src = ["."]
extend-exclude = ["**migrations**"]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "UP",  # pyupgrade
  "C90", # mccabe
  "I", # isort
  "N", # pep8-naming
  "ERA", # erradicate
  "PL", # pylint
  "TRY", # tryceratops 
  "FLY", # flynt 
]

[tool.pytest.ini_options]
pythonpath = "."
addopts = [
    "--import-mode=importlib",
]